Added Field::apply_traits
Field::apply_traits is a class method which injects a trait
into the Field base class. Can also be called on a derived
sub-class, in which case the trait will only apply to that
class and its descendents.

The performs the same job that the form-level field_traits
attribute previously performed.