Change to use Class::Load due to speed.
Class::MOP::load_class calls confess, and generating the strack trace
is quite expensive (especially if you're using Catalyst::Plugin::ErrorCatcher)
ergo switch over to using Class::Load, which provides almost the samebehaviour
but without the speed penalty.

Note that in _one_ case the handling was done correctly so that if a module
was present, but didn't compile then this error was detected. Using
Class::Load means that this situation will now be caught in _all_ cases.

Also swap the hand built type constraint for MooseX::Types::LoadableClass,
as it's neater and the latest version already uses Class::Load